a_i_t_backend/pyproject.toml
[project]
name = "a-i-t-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "exa-py>=1.15.6",
    "fastapi[standard]>=0.116.1",
    "google>=3.0.0",
    "google-adk>=1.13.0",
    "google-generativeai>=0.8.5",
    "google-search-results>=2.4.2",
    "mysql-connector-python>=9.4.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
]
